Update authorization dto and classes Removed redundant classes
