+++ src/SwedbankPay.Sdk.Infrastructure/PaymentInstruments/AuthorizationTransactionDto.cs
@@ -1,22 +1,29 @@
-﻿using SwedbankPay.Sdk.PaymentInstruments.Card;
-using System;
-using System.Collections.Generic;
+﻿using System;
 
 namespace SwedbankPay.Sdk.PaymentInstruments
 {
     internal class AuthorizationTransactionDto
     {
+        public long Amount { get; }
+        public long VatAmount { get; }
+        public DateTime Created { get; }
+        public string Description { get; }
+        public string FailedActivityName { get; }
+        public string FailedErrorCode { get; }
+        public string FailedErrorDescription { get; }
+        public string FailedReason { get; }
+        public bool IsOperational { get; }
+        public long Number { get; }
+        public OperationListDto Operations { get; }
+        public string PayeeReference { get; }
+        public string State { get; }
+        public string Type { get; }
+        public DateTime Updated { get; }
         public string Id { get; set; }
-        public List<PaymentAuthorizationDto> AuthorizationList { get; set; } = new List<PaymentAuthorizationDto>();
-        internal ICardPaymentAuthorizationListResponse Map()
+
+        internal IAuthorizationTransaction Map()
         {
-            var list = new List<IPaymentAuthorization>();
-            foreach (var item in AuthorizationList)
-            {
-                list.Add(new PaymentAuthorization(item));
-            }
-            var uri = new Uri(Id, UriKind.RelativeOrAbsolute);
-            return new CardPaymentAuthorizationListResponse(uri, list);
+            return new AuthorizationTransaction(this);
         }
     }
 }
