--- src/SwedbankPay.Sdk.Infrastructure/PaymentInstruments/Vipps/VippsPaymentAuthorizationTransactionDto.cs
@@ -1,45 +1,0 @@
-﻿using System;
-
-namespace SwedbankPay.Sdk.PaymentInstruments.Vipps
-{
-    internal class VippsPaymentAuthorizationTransactionDto
-    {
-        public string Id { get; set; }
-        public long Amount { get; set; }
-        public DateTime Created { get; set; }
-        public string Description { get; set; }
-        public string FailedActivityName { get; set; }
-        public string FailedErrorCode { get; set; }
-        public string FailedErrorDescription { get; set; }
-        public string FailedReason { get; set; }
-        public bool IsOperational { get; set; }
-        public long Number { get; set; }
-        public OperationListDto Operations { get; set; }
-        public string PayeeReference { get; set; }
-        public string State { get; set; }
-        public string Type { get; set; }
-        public DateTime Updated { get; set; }
-        public long VatAmount { get; set; }
-
-        internal AuthorizationTransaction Map()
-        {
-            var uri = new Uri(Id, UriKind.RelativeOrAbsolute);
-            return new AuthorizationTransaction(uri,
-                                                Created,
-                                                Updated,
-                                                Type,
-                                                State,
-                                                Number,
-                                                Amount,
-                                                VatAmount,
-                                                Description,
-                                                PayeeReference,
-                                                FailedReason,
-                                                FailedActivityName,
-                                                FailedErrorCode,
-                                                FailedErrorDescription,
-                                                IsOperational,
-                                                Operations.Map());
-        }
-    }
-}
